Revert "Merge branch 'feature/5108' into develop"
Reverts PR zendframework/zf2#5108

This reverts commit 5017af832b75d2bdf50f5ca26a8fb3c76782d9cd, reversing
changes made to c23c25099512244bb10773558310d5b6ec4de075.

The reason this is being reverted is because the solution does not work on
5.3.3; in fact, it will only work starting with version 5.3.9. Until we can
either up the minimum required PHP version or come up with an alternate
approach, this changeset cannot be included in a release.