Updated test for #19 to ensure it passes regardless of environment
The test as written did not work on my environment, and any attempt I
made to seed it with what should be a known-good value and set of
options failed. In the end, I used `IntlDateFormatter::create(/* ... */)->format(new DateTime())`,
per the basic validation tests, and this allowed it to work.